requirements.txt
django
channels
channels_redis
djangorestframework